### 2021-02-13 #429 @jabiologo
rWind: Download, Edit and Include Wind and Sea Currents Data in Ecological and Evolutionary Analysis
Submitting Author Name: Javier Fernández López
Submitting Author Github Handle: @jabiologo
Other Package Authors Github handles: (comma separated, delete if none) @KlausVigo, @PabloRL, Yurena Arjona
Repository: https://github.com/jabiologo/rWind
Version submitted:   1.1.6

Version accepted: TBD

---

```
Package: rWind
Title: Download, Edit and Include Wind and Sea Currents Data in Ecological and Evolutionary Analysis
Version: 1.1.6
Authors@R: c(person("Javier", "Fernández-López", email="jflopez.bio@gmail.com", role = c("aut", "cre")),
    person("Klaus", "Schliep", email="klaus.schliep@gmail.com", role = c("aut")),
    person("Yurena", "Arjona", email="yurenaaf@gmail.com", role = c("aut")))
Maintainer: Javier Fernández-López <jflopez.bio@gmail.com>
Description: Tools for download and manage surface wind and sea currents data from the Global Forecasting System <https://www.ncdc.noaa.gov/data-access/model-data/model-datasets/global-forcast-system-gfs> and to compute connectivity between locations.
License: GPL (>=3)
Imports: raster (>= 2.5-8), gdistance, Matrix, lubridate
Suggests: testthat, rmarkdown, knitr
Packaged: 2016-11-24 11:37:02 UTC; javi
NeedsCompilation: no
Repository: none
```

rWind contain tools for downloading, editing and transforming wind and sea current data from several sources. It also allows to use wind and sea currents data to compute the minimum cost path taking into account flow speed and direction to perform connectivity analysis.

Researchers interested in wind and sea currents data, specially focused in anisotropic connectivity analyses.

There are several packages to download and manage wind and sea currents data, but rWind allows to transform them in a easy way into a raster format and to compute connectivity cost analysis among different locations.
 rWind currently works with just two databases: wind data from Global Forecast System (GFS) of the USA's National Weather Service (NWS) (NOAA/NCEP Global Forecast System (GFS) Atmospheric Model; https://www.ncdc.noaa.gov/data-access/model-data/model-datasets/global-forcast-system-gfs), and ocean currents data from Ocean Surface Current Analyses Real-time (OSCAR) (https://coastwatch.pfeg.noaa.gov/erddap/info/jplOscar_LonPM180/index.html).
 rWind is specially designed for non physicists/meteorologists to provide an easy interface to download, manage and include wind data specially (but not only) in biological analyses (landscape genetics, dispersion, migration, etc.). After the data download, rWind provide easy utilities to transform raw data (U and V vector components from wind and ocean currents) in flow direction and speed or to compute time series averages of this data. After that, users can use rWind to transform those data into raster layers to be managed with raster packages or other Geographic Information System. With this raster layers, rWind can be used to compute wind/ocean currents mediated cost analysis between locations. rWind applies by default the formula described in Felicisimo et al. 2008, but other cost computations can be provided by the user in the flow.dispersion function. Thus, rWind provides a complete framework from time series data download, aggregate, manage and cost computation in a simple way for the users. It was published as a Software Note in the journal Ecography in 2019 (Fernández‐López & Schliep 2019) and it has been cited in at least 16 papers till now (https://scholar.google.es/scholar?oi=bibs&hl=es&cites=8441515588044324558).

There are other packages that covers data download and data managing that could overlap with rWind. However, rWind is specially focused to offer to the users a straightforward workflow from data download to cost analysis between locations. Some identified packages that cover partially the same utilities as rWind are:

rnoaa: Client for many 'NOAA' data sources including the 'NCDC' climate 'API', with functions foreach of the 'API' 'endpoints': data, data categories, data sets, data types,locations, location categories, and stations. This package allows the access of a huge NOAA data sources, but is not focused in transform data into raster format, nor in cost analyses. In addition, the high amount of data sources makes not easy for a non specialized user to decide what data base should be used for each analyses. It has not "official" vignettes at https://cran.r-project.org/web/packages/rnoaa/index.html

RNCEP: This package, similarly as the previous one, contains functions to retrieve, organize, and visualize weather data from the NCEP/NCAR Reanalysis, and NCEP/DOE Reanalysis II datasets. As the precious one, it is focused into obtain, manage (aggregate, etc) and visualize a huge amount of weather data. Visualization (NCEP.vis.area) are plot-based, non raster layers, and conversion among formats is not addressed in the package. Cost analyses are not implemented. It has not "official" vignettes at https://cran.r-project.org/web/packages/RNCEP/index.html. Kemp et al (2012)

weathercran: This package is developed to search for and download multiple months/years of historical weather data from Environment and Climate Change Canada (ECCC) website. Data bases are different to those used by rWind and data managing and cost analysis computations are not covered in this package. It has many "official" vignettes at https://cran.r-project.org/web/packages/weathercan/index.html.

In short, rWind is focused to fill the gap between wind and ocean currents data adquisition/managing and cost analysis among locations. It provides a straightforward framework to accomplish all these processes in a single R package.

References
Felicísimo, Á. M., Muñoz, J., & González-Solis, J. (2008). Ocean surface winds drive dynamics of transoceanic aerial movements. PLoS One, 3(8), e2928.
Fernández‐López, J., & Schliep, K. (2019). rWind: download, edit and include wind data in ecological and evolutionary analysis. Ecography, 42(4). https://onlinelibrary.wiley.com/doi/full/10.1111/ecog.03730 https://doi.org/10.1111/ecog.03730
Kemp, M. U., Emiel van Loon, E., Shamoun‐Baranes, J., & Bouten, W. (2012). RNCEP: global weather and climate data at your fingertips. Methods in Ecology and Evolution, 3(1), 65-70. https://besjournals.onlinelibrary.wiley.com/doi/10.1111/j.2041-210X.2011.00138.x

N/A
N/A
