### 2021-02-22 #429 @melvidoni
Hello @jabiologo, thanks for the submission.

Before we consider the package, we would like you to elaborate on the overlap of rWind with other packages. Please, edit your post and be as thorough as possible. Please let me know once it is done.
