### 2021-03-18 #429 @maurolepore
@jabiologo, here are the editor checks. Please respond to the comments prefixed with "(ml##)" and consider the other comments optional for now. Once you address the prefixed comments I'll start looking for reviewers.

### Editor checks:

- [x] **Automated tests:** Package has a testing suite and is tested via a CI service.

---

#### Editor comments

Documentation > README

* (ml01) Could you please clarify which are the data sources that rWind uses? The first
paragraph in README first refers to "several sources" then lists three specific
sources. And the first comment in software-review#429 refers to two (not three)
sources ("rWind currently works with just two databases ...").
* (ml02) Would you mind adding more details and context to the README? After reading it
someone with little domain knowledge should have been informed about the aim,
goals and functionality of the package.
* To benefit users, could you please add the most important examples directly to
readme -- instead of linking to an external page.
* README and/or vignettes is also where we recommend that packages highlight
differences from and improvements over overlapping packages.
* Consider removing the section "License" as the file DESCRIPTION
already includes license information.

Automated checks:

* (ml03) `spelling::spell_check_package()` finds a number of potentially misspelled
words. Please check them then run `spelling::update_wordlist()`.

``` r
spelling::spell_check_package()
#> DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
#>   WORD               FOUND IN
#> Á                 flow.dispersion_int.Rd:61
#>                    flow.dispersion.Rd:77
#> adviced            flow.dispersion_int.Rd:46
#>                    flow.dispersion.Rd:60
#> al                 flow.dispersion_int.Rd:14,35,42
#>                    flow.dispersion.Rd:28,49,56
#> allthiswasfield    rWind-package.Rd:17
#>                    wind.data.Rd:19
#>                    wind.series.Rd:20
#> anisotropic        flow.dispersion_int.Rd:27
#>                    flow.dispersion.Rd:41
#>                    Shortest_paths.Rmd:41
#> Anisotropic        Shortest_paths.Rmd:2
#> arrowDir           arrowDir.Rd:17,22,27
#> asimetric          flow.dispersion_int.Rd:28
#>                    flow.dispersion.Rd:42
#> betwen             flow.dispersion_int.Rd:25
#>                    flow.dispersion.Rd:39
#> blogspot           rWind-package.Rd:17
#>                    wind.data.Rd:19
#>                    wind.series.Rd:20
#> coastwatch         seaOscar.dl.Rd:47,70
#>                    README.md:17
#> codecov            README.md:8
#> colection          wind.dl_2.Rd:48
#>                    wind.dl.Rd:60
#> componenats        wind.mean.Rd:22
#> coordenate         seaOscar.dl.Rd:42
#>                    wind.dl_2.Rd:40,65
#>                    wind.dl.Rd:52
#> coordenates        wind.data.Rd:23
#>                    wind.dl_2.Rd:65
#>                    wind.dl.Rd:72
#> CRS                wind2raster_int.Rd:25
#>                    wind2raster.Rd:25
#> csv                seaOscar.dl.Rd:35,36,41,53,54
#>                    wind.dl_2.Rd:27,28,39,56,57
#>                    wind.dl.Rd:43,44,48,51,66,67
#> databse            seaOscar.dl.Rd:50
#> DCL                seaOscar.dl.Rd:68
#>                    wind.dl_2.Rd:88
#>                    wind.dl.Rd:86
#> dgCMatrix          flow.dispersion_int.Rd:17
#>                    flow.dispersion.Rd:31
#> diferent           flow.dispersion_int.Rd:16
#>                    flow.dispersion.Rd:30
#> dir                arrowDir.Rd:11
#>                    uv2ds.Rd:30
#>                    wind.dl_2.Rd:67
#>                    wind.dl.Rd:73
#>                    wind2raster_int.Rd:21
#> directionality     flow.dispersion_int.Rd:27
#>                    flow.dispersion.Rd:41
#> dl                 seaOscar.dl.Rd:35,46
#>                    wind.data.Rd:22
#>                    wind.dl_2.Rd:27,44,54,55,62
#>                    wind.dl.Rd:43,56
#>                    wind.fit_int.Rd:10,16,23
#>                    wind.series.Rd:23
#>                    wind2raster.Rd:21
#> doi                flow.dispersion_int.Rd:67
#>                    flow.dispersion.Rd:83
#> dplyr              tidy.rWind_series.Rd:19
#> ds                 uv2ds.Rd:22,24
#> edu                wind.data.Rd:37
#>                    wind.dl_2.Rd:51,90
#>                    wind.dl.Rd:63,88
#>                    wind.series.Rd:41
#> erddap             seaOscar.dl.Rd:47,70
#>                    wind.data.Rd:37
#>                    wind.dl_2.Rd:51,90
#>                    wind.dl.Rd:63,88
#>                    wind.series.Rd:41
#>                    README.md:17
#> et                 flow.dispersion_int.Rd:14,35,42
#>                    flow.dispersion.Rd:28,49,56
#> Etten              flow.dispersion_int.Rd:65
#>                    flow.dispersion.Rd:81
#> Felicísimo        flow.dispersion_int.Rd:14,42,61
#>                    flow.dispersion.Rd:28,56,77
#> Felícisimo        flow.dispersion_int.Rd:35
#>                    flow.dispersion.Rd:49
#> Fernández         arrowDir.Rd:56
#>                    flow.dispersion_int.Rd:73
#>                    flow.dispersion.Rd:89
#>                    rWind-package.Rd:20,26
#>                    seaOscar.dl.Rd:76
#>                    uv2ds.Rd:42
#>                    wind.dl_2.Rd:97
#>                    wind.dl.Rd:94
#>                    wind.fit_int.Rd:33
#>                    wind.mean.Rd:39
#>                    wind2raster_int.Rd:32
#>                    wind2raster.Rd:41
#> forcast            wind.dl_2.Rd:46
#>                    wind.dl.Rd:58
#>                    README.md:13
#> gdistance          flow.dispersion_int.Rd:19,29,65
#>                    flow.dispersion.Rd:33,43,81
#> Geospatial         seaOscar.dl.Rd:48
#>                    wind.dl_2.Rd:48
#>                    wind.dl.Rd:60
#> gfs                wind.dl_2.Rd:46
#>                    wind.dl.Rd:58
#>                    README.md:13
#> GFS                rWind-package.Rd:10
#>                    wind.dl_2.Rd:15,45,47,66
#>                    wind.dl.Rd:30,56,59,72
#>                    wind.fit_int.Rd:17,19
#> ggplot             tidy.rWind_series.Rd:19
#> ggvis              tidy.rWind_series.Rd:19
#> gis                seaOscar.dl.Rd:68
#>                    wind.dl_2.Rd:88
#>                    wind.dl.Rd:86
#> gmail              rWind-package.Rd:26
#>                    seaOscar.dl.Rd:76
#>                    uv2ds.Rd:42
#>                    wind.dl_2.Rd:97
#>                    wind.dl.Rd:94
#>                    wind.fit_int.Rd:33
#>                    wind.mean.Rd:39
#>                    wind2raster_int.Rd:32
#>                    wind2raster.Rd:41
#> González          flow.dispersion_int.Rd:61
#>                    flow.dispersion.Rd:77
#> griddap            wind.dl_2.Rd:90
#>                    wind.dl.Rd:63,88
#> gridded            wind2raster_int.Rd:17
#>                    wind2raster.Rd:17
#> hawaii             wind.data.Rd:37
#>                    wind.dl_2.Rd:51,90
#>                    wind.dl.Rd:63,88
#>                    wind.series.Rd:41
#> HorizontalFactor   flow.dispersion_int.Rd:39,41
#>                    flow.dispersion.Rd:53,55
#> http               rWind-package.Rd:17
#>                    seaOscar.dl.Rd:68
#>                    wind.data.Rd:19,37
#>                    wind.dl_2.Rd:51,88,90
#>                    wind.dl.Rd:86
#>                    wind.series.Rd:20,41
#>                    README.md:15
#> https              arrowDir.Rd:50
#>                    seaOscar.dl.Rd:47,70
#>                    wind.dl_2.Rd:46
#>                    wind.dl.Rd:58,63,88
#>                    wind.fit_int.Rd:26
#>                    wind.mean.Rd:33
#>                    README.md:13,17
#> incrementaly       flow.dispersion_int.Rd:41
#>                    flow.dispersion.Rd:55
#> internaly          wind.dl_2.Rd:66
#>                    wind.dl.Rd:73
#>                    wind.fit_int.Rd:16,23
#> jflopez            rWind-package.Rd:26
#>                    seaOscar.dl.Rd:76
#>                    uv2ds.Rd:42
#>                    wind.dl_2.Rd:97
#>                    wind.dl.Rd:94
#>                    wind.fit_int.Rd:33
#>                    wind.mean.Rd:39
#>                    wind2raster_int.Rd:32
#>                    wind2raster.Rd:41
#> jplOscar           seaOscar.dl.Rd:47,70
#>                    README.md:17
#> jss                flow.dispersion_int.Rd:67
#>                    flow.dispersion.Rd:83
#> Karline            arrowDir.Rd:48
#> logitude           wind2raster_int.Rd:18
#>                    wind2raster.Rd:19
#> lon                seaOscar.dl.Rd:43
#>                    wind.dl_2.Rd:41
#>                    wind.dl.Rd:53
#> LonPM              seaOscar.dl.Rd:47,70
#>                    README.md:17
#> López             arrowDir.Rd:56
#>                    flow.dispersion_int.Rd:73
#>                    flow.dispersion.Rd:89
#>                    rWind-package.Rd:20,26
#>                    seaOscar.dl.Rd:76
#>                    uv2ds.Rd:42
#>                    wind.dl_2.Rd:97
#>                    wind.dl.Rd:94
#>                    wind.fit_int.Rd:33
#>                    wind.mean.Rd:39
#>                    wind2raster_int.Rd:32
#>                    wind2raster.Rd:41
#> Muñoz             flow.dispersion_int.Rd:61
#>                    flow.dispersion.Rd:77
#> ncdc               wind.dl_2.Rd:46
#>                    wind.dl.Rd:58
#>                    README.md:13
#> ncep               wind.dl.Rd:63,88
#> NCEP               wind.data.Rd:37
#>                    wind.dl_2.Rd:47,51,90
#>                    wind.dl.Rd:59
#>                    wind.series.Rd:41
#> noaa               seaOscar.dl.Rd:47,70
#>                    wind.dl_2.Rd:46
#>                    wind.dl.Rd:58
#>                    README.md:13,17
#> NOAA               wind.dl_2.Rd:47
#>                    wind.dl.Rd:59
#> nocite             Shortest_paths.Rmd:14
#> NWS                wind.dl_2.Rd:45
#>                    wind.dl.Rd:57
#> oos                wind.data.Rd:37
#>                    wind.dl_2.Rd:51,90
#>                    wind.series.Rd:41
#> oscar              seaOscar.dl.Rd:36
#> pacioos            wind.dl.Rd:63,88
#> pae                wind.dl.Rd:63,88
#> paha               wind.dl.Rd:63,88
#> pfeg               seaOscar.dl.Rd:47,70
#>                    README.md:17
#> PLoS               flow.dispersion_int.Rd:62
#>                    flow.dispersion.Rd:78
#> POSIXt             wind.dl_2.Rd:13
#> procesed           uv2ds.Rd:30
#> rasterFromXYZ      wind2raster_int.Rd:19
#>                    wind2raster.Rd:20
#> RasterStack        flow.dispersion_int.Rd:10
#>                    flow.dispersion.Rd:24
#> rwind              wind.data.Rd:19
#>                    wind.series.Rd:20
#> seaOscar           seaOscar.dl.Rd:35,46
#> skalar             flow.dispersion.Rd:13,15
#> soest              wind.data.Rd:37
#>                    wind.dl_2.Rd:51,90
#>                    wind.series.Rd:41
#> Soetaert           arrowDir.Rd:48
#> tendences          wind.dl_2.Rd:64
#> th                 Shortest_paths.Rmd:46
#> transfored         wind.dl_2.Rd:14
#> transitionLayer    flow.dispersion_int.Rd:18,24
#>                    flow.dispersion.Rd:32,38
#> TransitionLayer    flow.dispersion_int.Rd:18,29
#>                    flow.dispersion.Rd:32,43
#> tt                 wind.dl_2.Rd:28,57
#>                    wind.dl.Rd:44,67
#> uv                 uv2ds.Rd:22,24
#> WDOWmbV            seaOscar.dl.Rd:68
#>                    wind.dl_2.Rd:88
#>                    wind.dl.Rd:86
#> WGS                wind2raster_int.Rd:25
#>                    wind2raster.Rd:25
#> wikipedia          wind.fit_int.Rd:26
#>                    wind.mean.Rd:33
#> www                seaOscar.dl.Rd:68
#>                    wind.dl_2.Rd:46,88
#>                    wind.dl.Rd:58,86
#>                    README.md:13,15
#> yyyy               seaOscar.dl.Rd:36,54
#>                    wind.dl_2.Rd:28,57
#>                    wind.dl.Rd:44,67
```

* (ml04) `goodpractice::gp()` finds a number of issues. Please try to address or
justify them in the particular context of this package.

``` r
goodpractice::gp()
#> Preparing: covr
#> Preparing: cyclocomp
#>
#>      checking for file ‘/tmp/Rtmp3bi5u3/remotes96556a3cb92/rWind/DESCRIPTION’ ...  ✓  checking for file ‘/tmp/Rtmp3bi5u3/remotes96556a3cb92/rWind/DESCRIPTION’
#>   ─  preparing ‘rWind’:
#> ✓  checking DESCRIPTION meta-information
#>      checking vignette meta-information ...  ✓  checking vignette meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#>   ─  building ‘rWind_1.1.6.tar.gz’
#>
#>
#> Preparing: description
#> Preparing: lintr
#> Preparing: namespace
#> Preparing: rcmdcheck
#> ── GP rWind ────────────────────────────────────────────────────────────────────
#>
#> It is good practice to
#>
#>   ✖ write unit tests for all functions, and all package code in
#>     general. 36% of code lines are covered by test cases.
#>
#>     R/wind_functions2.R:81:NA
#>     R/wind_functions2.R:83:NA
#>     R/wind_functions2.R:84:NA
#>     R/wind_functions2.R:85:NA
#>     R/wind_functions2.R:88:NA
#>     ... and 232 more lines
#>
#>   ✖ avoid long code lines, it is bad for readability. Also, many people
#>     prefer editor windows that are about 80 characters wide. Try make
#>     your lines shorter than 80 characters
#>
#>     R/rWind-package.R:3:1
#>     R/rWind-package.R:5:1
#>     R/rWind-package.R:6:1
#>     R/wind_functions2.R:27:1
#>     R/wind_functions2.R:115:1
#>     ... and 31 more lines
#>
#> ────────────────────────────────────────────────────────────────────────────────
```

* The file .travis.yml suggests the continuous integration service that rWind
uses is Travis. [We recommend moving away from
Travis](https://devguide.ropensci.org/ci.html#travis-ci-linux-and-mac-osx) and
instead using GitHub actions.

* The file .travis.yml suggests the package is tested only on a single platform.
[If applicable](https://devguide.ropensci.org/ci.html#whichci), consider testing
with the three main platforms.

* (ml05) The file .travis.yml suggests the package is tested only with
R-release. We require that rOpenSci packages are tested against the latest,
previous and development versions of R to ensure both backwards and forwards
compatibility with base R.

* Consider running an automated code-styler such as `usethis::use_tidy_style()`.
This standard makes the code more readable for reviewers and contributors.

* I expected a single .Rproj file but instead I see two:

``` bash
# bash
ls -l *.Rproj
```

    $ -rwxrwxr-x 1 rstudio rstudio 356 Mar 13 08:29 rWind1.Rproj
    $ -rw-rw-r-- 1 rstudio rstudio 343 Mar 13 20:34 rWind.Rproj

-   DESCRIPTION

* `urlchecker::urlcheck()` finds problematic URLs. You may fix them with
`urlchecker::url_update()`.

* Consider running `usethis::use_tidy_description()` to organize DESCRIPTION in
a standard way and facilitate review.

* These fields are not included when running
`usethis::create_package()`. Are they necessary? The field
"Packaged" seems outdated; for maintainability consider removing
it.

        -NeedsCompilation: no
        -Packaged: 2016-11-24 11:37:02 UTC; javi
        -Repository: none

    -   With `usethis::create_package()` the value of `LazyData` is
        lowercase `true` not uppercase `TRUE`. Please check it's the
        same or better use the `true` to stick to usethis default.

* [We strongly recommend snake_case over all other styles unless you are porting
over a package that is already in wide
use](https://devguide.ropensci.org/building.html?q=snake#function-and-argument-naming
).

* Consider using the convention by which every exported function `f()` should
have two corresponding files named "R/f.R" and "tests/testthat/test-f.R". This
convention plays well with development tools from the usethis package (e.g.
`use_r()` and `use_test()`) and the testthat package (as test failures report
the name of the problematic test-file).

* Consider using rmarkdown syntax in roxygen documentation. You may enable it
with `usethis::use_roxygen_md()`. This makes roxygen documentation easier to
maintain and read.

---

Reviewers: TBC
Due date: TBC
