### 2021-03-19 08c7365
remove travisCI, fix 1 url
