### 2021-03-19 2c04a24
add github actions for checks
