+++ R/wind_functions2.R
@@ -22,7 +22,7 @@ circ.mean <- function(deg){
 #' USA's National Weather Service (NWS)
 #' (https://www.ncdc.noaa.gov/data-access/model-data/model-datasets/global-forcast-system-gfs).
 #' Wind data are taken from NOAA/NCEP Global Forecast System (GFS) Atmospheric
-#' Model colection. Geospatial resolution is 0.5 degrees (approximately 50 km),
+#' Model collection. Geospatial resolution is 0.5 degrees (approximately 50 km),
 #' and wind is calculated for Earth surface, at 10 m. More metadata
 #' information:
 #' https://pae-paha.pacioos.hawaii.edu/erddap/griddap/ncep_global.graph
@@ -48,10 +48,10 @@ circ.mean <- function(deg){
 #' @param trace if trace = 1 (by default) track downloaded files
 #' @param file file name of the saved ".csv" files.
 #' @return "rWind" and "data.frame" class object or .csv file with U and V
-#' vector  components and wind direction and speed for each coordenate
+#' vector  components and wind direction and speed for each coordinate
 #' in the study area defined by lon1/lon2 and lat1/lat2.
-#' @note Longitude coordenates are provided by GFS dataset in 0/360 notation
-#' and transformed internaly into -180/180. Wind "dir" denotes where the
+#' @note Longitude coordinate are provided by GFS dataset in 0/360 notation
+#' and transformed internally into -180/180. Wind "dir" denotes where the
 #' wind is going (toward), not from where is coming.
 #' @author Javier Fernández-López (jflopez.bio@@gmail.com)
 #' @seealso \code{\link{wind.dl_2}}, \code{\link{wind2raster}}
@@ -173,7 +173,7 @@ read.rWind <- function(file){
 #' an \code{rWind_series} object is created.
 #'
 #' @param time a scalar or vector of POSIXt or Date objects or an character
-#' which can transfored into those, see example below.
+#' which can transformed into those, see example below.
 #' There are currently these options at the GFS database for the hours:
 #' 00:00 - 03:00 - 06:00 - 09:00 - 12:00 - 15:00 - 18:00 - 21:00 (UTC) (TO).
 #' @param lon1 Western longitude
@@ -185,13 +185,13 @@ read.rWind <- function(file){
 #' directory named "wind_yyyy_mm_dd_tt.csv".
 #' @param trace if trace = 1 (by default) track downloaded files
 #' @return an object of class \code{rWind_series} or .csv file/s with
-#' U and V vector components and wind direction and speed for each coordenate
+#' U and V vector components and wind direction and speed for each coordinate
 #' in the study area defined by lon1/lon2 and lat1/lat2.
 #' @note wind.dl_2 requires two dates that represent the boundaries of the time
 #' lapse to download wind series data.
-#' U and V vector components allow you to create wind averages or tendences
-#' for each coordenate at the study area. Longitude coordenates are
-#' provided by GFS dataset in 0/360 notation and transformed internaly into
+#' U and V vector components allow you to create wind averages or tendencies
+#' for each coordinate at the study area. Longitude coordinates are
+#' provided by GFS dataset in 0/360 notation and transformed internally into
 #' -180/180. "dir" denotes where the
 #' wind/sea current is going (toward), not from where is coming.
 #' @author Javier Fernández-López (jflopez.bio@@gmail.com)
@@ -268,7 +268,7 @@ wind.dl_2 <- function(time, lon1, lon2, lat1, lat2, type="read-data", trace=1){
                     write.table(tmp, fname, sep = ",", row.names = FALSE,
                                 col.names = TRUE, quote = FALSE)
                 }
-                else{
+                else{coorde
                     resultados[[id]] <- tmp[, 4:5]
                 }
             }
[…]
