### 2021-04-06 #429 @jabiologo
Hi @maurolepore thanks for contacting. My colleague Klaus already did some changes:

    -> remove travisCI, fix 1 url
    -> add github actions for checks
    -> remove redundant file
    -> fix some spelling mistakes
    -> clean up DESCRIPTION file
    -> bug fix
    -> add test again

but I was too busy with some deadlines and I'd need more time to finish some more things and respond the comments. Could I have one more week? Thanks in advance.
