### 2021-04-27 5e8ca27
updating README for rOpenScience
