+++ README.md
@@ -8,22 +8,111 @@
 [![codecov](https://codecov.io/gh/jabiologo/rWind/branch/master/graph/badge.svg)](https://codecov.io/gh/jabiologo/rWind)
 
 ## Overview
-rWind contain tools for downloading, editing and transforming wind and sea current data from several sources. It also allows to use wind and sea currents data to compute the minimum cost path taking into account flow speed and direction to perform connectivity analysis. For more information about data source, please check: 
 
-* <https://www.ncdc.noaa.gov/data-access/model-data/model-datasets/global-forcast-system-gfs>
+ rWind is a library in the R language for statistical computing and graphics (R Development Core Team), designed specifically to download and process wind and sea currents data from the Global Forecasting System. From these data, users can obtain wind/sea currents speed and direction layers in order to compute connectivity values between locations. There are other great R libraries that covers data download and data managing that could overlap with rWind, such us [rnoaa](https://cran.r-project.org/web/packages/rnoaa/index.html), [RNCEP](https://cran.r-project.org/web/packages/RNCEP/index.html) or [weathercran](https://cran.r-project.org/web/packages/weathercan/index.html). However, rWind is specially focused to offer to the users a straightforward workflow from data download to cost analysis between locations. rWind fills the gap between wind/sea currents data accessibility and their inclusion in a general framework to be applied broadly in ecological or evolutionary studies.
+ 
+ It has been peer-reviewed published in: Fernández‐López, J. and Schliep, K. (2019), rWind: download, edit and include wind data in ecological and evolutionary analysis. Ecography, 42: 804-810. https://doi.org/10.1111/ecog.03730  
+https://onlinelibrary.wiley.com/doi/full/10.1111/ecog.03730
 
-* <http://www.weather.gov/>
+ 
 
-* <https://coastwatch.pfeg.noaa.gov/erddap/info/jplOscar_LonPM180/index.html>
+ For more information about data source, please check: 
+
+NOAA/NCEP Global Forecast System (GFS) Atmospheric Model colection (wind data)  
+* <https://pae-paha.pacioos.hawaii.edu/erddap/info/ncep_global/index.html>
+
+Ocean Surface Current Analyses Real-time (OSCAR) (sea currents data)  
+* <https://coastwatch.pfeg.noaa.gov/erddap/info/jplOscar_LonPM180/index.html><br />
+
+To install the latest released version of rWind on CRAN use `install.packages("rWind")`  
+To install the latest development version `devtools::install_github("jabiologo/rWind")`  
+  
+  
+  
+### Quick example: Computing anisotropic shortest paths across Strait of Gibraltar
+
+First, we load the main packages we will use in this example. This vignette 
+was written under the rWind version `r foo$Version`
+
+```{R}
+# use install.packages() if some is not installed
+# you can install the latest development version using the command 
+# devtools::install_github("jabiologo/rWind")
+library(rWind)
+library(raster)
+library(gdistance)
+```
+
+
+In this simple example, we introduce the most basic functionality of rWind, 
+to get the shortest paths between two points across Strait of Gibraltar. Notice
+that, as wind connectivity is anisotropic (direction dependent), shortest path
+from A to B usually does not match with shortest path from B to A.
+
+First, we download wind data of a selected date (e.g. 2015 February 12th). 
+
+```{R}
+w <- wind.dl(2015, 2, 12, 12, -7, -4, 34.5, 37.5)
+```
[…]
