+++ R/wind_functions2.R
@@ -1,18 +1,22 @@
 # some trigonemetric functions
-rad2deg <- function(rad) {(rad * 180) / (pi)}
-deg2rad <- function(deg) {(deg * pi) / (180)}
+rad2deg <- function(rad) {
+  (rad * 180) / (pi)
+}
+deg2rad <- function(deg) {
+  (deg * pi) / (180)
+}
 
 # circular mean
 # https://en.wikipedia.org/wiki/Mean_of_circular_quantities
-circ.mean <- function(deg){
-    rad.m <- (deg * pi) / (180)
-    mean.cos <- mean(cos(rad.m))
-    mean.sin <- mean(sin(rad.m))
+circ.mean <- function(deg) {
+  rad.m <- (deg * pi) / (180)
+  mean.cos <- mean(cos(rad.m))
+  mean.sin <- mean(sin(rad.m))
 
-    theta <- rad2deg(atan(mean.sin/mean.cos))
-    if(mean.cos < 0) theta <- theta + 180
-    if((mean.sin < 0) & (mean.cos > 0)) theta <- theta + 360
-    theta
+  theta <- rad2deg(atan(mean.sin / mean.cos))
+  if (mean.cos < 0) theta <- theta + 180
+  if ((mean.sin < 0) & (mean.cos > 0)) theta <- theta + 360
+  theta
 }
 
 
@@ -65,91 +69,109 @@ circ.mean <- function(deg){
 #' # Download wind for Iberian Peninsula region at 2015, February 12, 00:00
 #' \dontrun{
 #'
-#' wind.dl(2015,2,12,0,-10,5,35,45)
-#'
+#' wind.dl(2015, 2, 12, 0, -10, 5, 35, 45)
 #' }
-#'
 #'
 #' @importFrom utils write.table read.csv download.file
 #' @importFrom lubridate ymd_h year month day hour
 #' @rdname wind.dl
 #' @export wind.dl
-wind.dl <- function (yyyy,mm,dd,tt,lon1,lon2,lat1,lat2,
-                         type="read-data", trace=1){
+wind.dl <- function(yyyy, mm, dd, tt, lon1, lon2, lat1, lat2,
+                    type = "read-data", trace = 1) {
+  type <- match.arg(type, c("read-data", "csv"))
 
-    type <- match.arg(type, c("read-data", "csv"))
+  mm <- sprintf("%02d", mm)
+  dd <- sprintf("%02d", dd)
+  tt <- sprintf("%02d", tt)
 
-    mm<-sprintf("%02d", mm)
-    dd<-sprintf("%02d", dd)
-    tt<-sprintf("%02d", tt)
+  # Create a sequence with all dates available between selected dates
+  dt <- ymd_h(paste(yyyy, mm, dd, tt, sep = "-"))
 
-    # Create a sequence with all dates available between selected dates
-    dt <- ymd_h(paste(yyyy,mm,dd,tt, sep="-"))
+  yyyy_c <- year(dt)
+  mm_c <- sprintf("%02d", month(dt))
+  dd_c <- sprintf("%02d", day(dt))
+  tt_c <- sprintf("%02d", hour(dt))
 
-    yyyy_c <- year(dt)
-    mm_c <- sprintf("%02d",month(dt))
-    dd_c <- sprintf("%02d",day(dt))
-    tt_c <- sprintf("%02d",hour(dt))
+  testDate <- paste(yyyy_c, "-", mm_c, "-", dd_c, sep = "")
+  print(testDate)
+  if (trace) {
+    print(paste(ymd_h(paste(yyyy_c, mm_c, dd_c, tt_c, sep = "-")),
+      "downloading...",
+      sep = " "
+    ))
+  }
 
[…]
