+++ tests/testthat/test_rWind.R
@@ -1,7 +1,9 @@
 context("test rWind")
 
-X <- data.frame("2017-01-01T00:00:00Z", rep(c(-0.5,0,0.5), each=3),
-                rep(c(1,1.5,2)), 1:9, 9:1, 1:9, 9:1)
+X <- data.frame(
+  "2017-01-01T00:00:00Z", rep(c(-0.5, 0, 0.5), each = 3),
+  rep(c(1, 1.5, 2)), 1:9, 9:1, 1:9, 9:1
+)
 colnames(X) <- c("time", "lat", "lon", "ugrd10m", "vgrd10m", "dir", "speed")
 class(X) <- c("rWind", "data.frame")
 
@@ -31,20 +33,19 @@ test_that("rWind works as expected", {
 
 
 test_that("reading files works as expected", {
-    tmp <- tempfile()
-    write.csv(wind.data, file = tmp, row.names = FALSE)
-    tmp2 <- read.rWind(tmp)
-    unlink(tmp)
-    expect_equal(wind.data, tmp2, check.attributes = FALSE)
+  tmp <- tempfile()
+  write.csv(wind.data, file = tmp, row.names = FALSE)
+  tmp2 <- read.rWind(tmp)
+  unlink(tmp)
+  expect_equal(wind.data, tmp2, check.attributes = FALSE)
 })
 
 
 # may works in future testthat version from https://github.com/r-lib/testthat
 test_that("downloading works", {
-    skip_if_offline()
-    skip_on_cran()
-    dl1 <- wind.dl(2015,2,12,0,-10,5,35,45)
-    dl2 <- wind.dl_2("2015/2/12 0:00:00",-10,5,35,45)
-    expect_equal(dl1, dl2[[1]])
+  skip_if_offline()
+  skip_on_cran()
+  dl1 <- wind.dl(2015, 2, 12, 0, -10, 5, 35, 45)
+  dl2 <- wind.dl_2("2015/2/12 0:00:00", -10, 5, 35, 45)
+  expect_equal(dl1, dl2[[1]])
 })
-
