### 2021-04-27 #429 @jabiologo
Hi @maurolepore, sorry for the delay and thanks for your constructive commentaries. We addressed the most of the suggestion you made, please find below a detailed list with some responses:

### Documentation > README

(**ml01**) rWind currently works with just two databases, one for wind data and other for sea currents data, one of the URLs was duplicated since it referred to the same database. We have corrected this.

(**ml02**) We have added more detailed description of the package as well as the reference to the peer reviewed published paper about rWind for more information.

**.** We have added one simple example from the vignette into the README file.

**.** We have added some statements about differences and improvements over other packages.

**.** We have removed the “License” section

### Automated checks:

(**ml03**) Spell check has been conducted and word list has been updated
(**ml04**) goodpractice::gc() has been run. It suggested two main issues: only 41% of our code is currently tested. Current time availability doesn’t allow us to afford new test… The other main issue was related to too long code lines. Those lines are the URLs used by rWind to connect to different data bases to download wind and sea currents data, so we consider it makes sense to maintain as they are.
**.** Travis file has been removed and GitHub actions are now used
**.** The package tests are now conducted in Windows, iOS and Linux
(**ml05**) GitHub actions are now used to test the package
**.** use_tidy_style() done and updated
**.** rWind1.Rproj has been removed

### DESCRIPTION

**.**urlchecker::url_update() has been run. It reports an issue related to https://www.ncdc.noaa.gov/data-access/model-data/model-datasets/global-forcast-system-gfs. However, we think it is due to the way the original server is managing the query for this URL.
**.**usethis issues have been addressed
**.**snake_case issue: rWind is already in wide use, the paper has been cited at least  19 times in other peer reviewed papers and many blog post and tweets have been cited. In any case, we can consider in a near future to make something similar as the package igraph, in which they maintain both, original and snake_case functions. Could this work for you?
**.**We’ll consider the "R/f.R" and "tests/testthat/test-f.R" structure in near future.
**.**usethis::use_roxygen_md() done

Thanks again for your work.
