### 2021-04-28 #429 @maurolepore
@jabiologo,

Thanks for addressing most of my most pressing concerns.

Our guide states that _[Test coverage below 75% will likely require additional tests or explanation before being sent for review](https://devguide.ropensci.org/building.html#testing)_. Although__ test coverage near 100% is a poor predictor of high-quality code, low test coverage is a good predictor of poor quality code ([Unit testing principles, practices, and patterns](https://www.amazon.com/-/es/Vladimir-Khorikov/dp/1617296279)).

Increasing test coverage might not be too time consuming and will likely pay off. If appropriate you may [exclude specific chunks of code](https://covr.r-lib.org/#exclusions) with an explanatory comment. And if you haven't seen them yet, check [rOpenSci's resources on testing R packages](https://twitter.com/rOpenSci/status/1386749574480465923?s=20), and let me know if you get stuck.

I totally understand you are busy. It happens and you should not worry. Would you like to give it a go or prefer to hold?

(The hold has no impact other than freeing me to help review other packages)
