### 2021-08-01 #429 @maurolepore
Hi @jabiologo, it's about the time when we should review the "holding" status. Shall we keep it?

--
> The author can choose to have their submission put on hold (editor applies the holding label). The holding status will be revisited every 3 months, and after one year the issue will be closed. --
-- https://devdevguide.netlify.app/policies.html#policiesreviewprocess
