+++ man/uv2ds.Rd
@@ -32,10 +32,8 @@ wind/sea current is going (toward), not from where is coming.
 }
 \examples{
 
-( ds <- uv2ds(c(1,1,3,1), c(1,1.7,3,1)) )
-ds2uv(ds[,1], ds[,2])
-
-
+(ds <- uv2ds(c(1, 1, 3, 1), c(1, 1.7, 3, 1)))
+ds2uv(ds[, 1], ds[, 2])
 }
 \seealso{
 \code{\link{wind.mean}}, \code{\link{wind2raster}}
