+++ man/wind.dl.Rd
@@ -78,10 +78,8 @@ wind is going (toward), not from where is coming.
 # Download wind for Iberian Peninsula region at 2015, February 12, 00:00
 \dontrun{
 
-wind.dl(2015,2,12,0,-10,5,35,45)
-
+wind.dl(2015, 2, 12, 0, -10, 5, 35, 45)
 }
-
 
 }
 \references{
