### 2021-10-15 3669a37
some fix in arrowDir documentation
