+++ man/arrowDir.Rd
@@ -19,7 +19,7 @@ arrowDir adapts wind direction value to be used by Arrowhead function from
 }
 \details{
 Angle argument of Arrowhead function from "shape" package needs to be fed
-in an anti-clockwise way, relative to x-axis, in degrees \link{0,360}. arrowDir
+in an anti-clockwise way, relative to x-axis, in degrees 0,360. arrowDir
 function adapts wind direction provided by wind.fit (clockwise, relative to
 y-axis ) to requirements of Arrowhead.
 }
