### 2021-10-15 ac0832d
updating canonical form of URLs
