### 2021-12-20 f82f0a3
Update object to use in image.plot
From 'wind_layer$wind.speed' to 'wind_layer[["speed"]]'
