### 2022-01-24 b428ef3
Allowing only passive cost computation
