### 2022-01-31 #429 @maurolepore
Hi @jabiologo, it's time to remind you that this package continues on hold. Let me know whenever you're ready to proceed with the review. Else I'll close this issue on April 30, 2022.

--

The author can choose to have their submission put on hold (editor applies the holding label). The holding status will be revisited every 3 months, and after one year the issue will be closed.
-- https://devdevguide.netlify.app/policies.html#policiesreviewprocess
