### 2022-05-30 #429 @maurolepore
Hi @jabiologo, I'm closing this issue now because it's been on hold for more than 1 year.

--

The author can choose to have their submission put on hold (editor applies the holding label). The holding status will be revisited every 3 months, and after one year the issue will be closed.
-- https://devdevguide.netlify.app/policies.html#policiesreviewprocess
